backend/requirements.txt
anyio==4.8.0
asyncpg==0.30.0
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
pyjwt
SQLAlchemy==2.0.38
greenlet==3.1.1